backend/requirements.txt
djangorestframework==3.15.2
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1requests==2.32.3
